reduced warnings
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@21543 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ asl/base/testMappedBlock.tst.cpp
@@ -255,10 +255,10 @@ public:
             // this is bad if e.g. while downloading the program is interrupted
             // with CTRL-C since there will then be garbage behind the data.
             // reserve seems to be very expensive on win32 with large files
-            ENSURE(getFileSize(testFileName) == myData1.size());   
+            ENSURE( static_cast<unsigned long>(getFileSize(testFileName)) == myData1.size() );
         }
         // after closing the size is ok.
-        ENSURE(getFileSize(testFileName) == myData1.size());   
+        ENSURE( static_cast<unsigned long>(getFileSize(testFileName)) == myData1.size() );
         {
             MappedBlock myBlock(testFileName); // this should work
             // MappedBlock myBlock(testFileName, myData1.size()); // but we need this
